requirements.txt
requests
python-dotenv
openai
pydantic
